README.md + S5P3T0
I wrote a complete README.md by hand holding all the informations and course I gathered on argc argv  and unused variable, which is very useful. Then I did the first (0) task which was simply to use argc and argv to print the name of the program. I just had to choose a way to ignore argc here. I chose (void)argc for a clear code.