S5P3T4 - This task has been quite difficult to wrap my head around it. I managed to grasp the whole idea but need the consult man and help for proper use of isdigit and atoi.
In this task I had to program a script to addition each positiver number and sent an error if something else. Considering that each argv could have more than 1 characters. The whole script was ok to do but I had issue to know how to verify each character in EACH argv. I asked ChatGPT for help for this and it was quite easy in fact. Just using another favriable to creat a second index to loop as any other to cheack with "isdigit" each character from each argv.