S6P1T2 - In this tasl I complicated a lot myself while in fact, it was working. So I had to concatenate 2 strings with proper allowance of space with malloc.
For that I counted with a loop what their lengh for (next time I will try to use _strlen)
Then I used malloc on a new array of char to allocate the proper amount of space. And then I used twice strncat to concatenate in my new array with the proper size.