S6P3T1 - This task was to initialized an instance of "struct dog" with a new prototype.
In itself it wasn't difficult BUT I had  to think about my headers. 1-main.c didn't hold and #include main.h which suggest that the checker won't too. In this case then I have to include my prototype definition in my header "dog.h"
On the same note, I had to think about adding it to the gcc compiler line.